split classes out of "main" gtp2ogs binary
This change splits the monolithic gtp2ogs.js file into separate files.
The motivation for this is to be able to write tests for gtp2ogs. For
tests, we need to load parts of the functionality (ex: the Bot class) at
a time and not trigger a full run of the application.

This change is large but largely mechanical:

1. Each class is split into a separate file with the util functions used
   by that class (with a few exceptions for helpers used by multiple
   classes).
2. All the global configuration variables (arising from command line
   arguments) are moved into a config.js.
3. Class code is updated to refer to CONFIG variables (ex: DEBUG) and
   argv through config\. reference.

There were a few tiny changes that were not mechanical. Since they might
get lost in the noise of the mechanical changes otherwise, I'm pointing
them out separately:

1. (corr_)_moves_processing is now a "static" variable on Game class.
2. Updated a few global "conn" references in Connection code to refer to
   "this" instead.