avoid starving correspondence game queue
I had observed some games with correspondence time settings but players
playing them as live games. This is actually a decent way to play bot
games - you can play a live game but walk away any time something
happens in real life. The bot won't mind the wait.

However, if several people do this, then it happens that whoevers' game
is in the beginning of the game list will tend to get their moves
responded to quickly and whoever happens to be at the back of the list
will have to wait a long time - games will be left waiting for the bot
to move for potentially hours.

I think technically this could cause bot games to time out but that is
unlikely to happen in practice as it would require constantly active
correspondence queues 24 hours of the day. So the result is just that
some players will suddenly find that the correspondence game suddenly
turns into an actual proper correspondence game where you have to wait
for the opponent to move.

Despite being technically OK, this is annoying to the users. I have
gotten some questions asked about this and there have been some game
cancellatins when this happens early in the game.

This change adds a test case that demonstrates the undesirable behavior
and fixes the test by picking a random correspondence game to make a
move for, rather than picking the first one in the list.

I also changed the interval for considering this from 10s to 1s - it
should make hardly any difference from gtp2ogs side but should increase
the interactivity of correspondence games by a fair bit and keep at
least one game going most of the time.