rework game timeouts
Currently game timeouts can be triggered both during bot turn and
opponent turn. If they are triggered during player turn, then the game
is forgotten (for the time being) and can only be recovered when the
server sends another message.

In practice, it seems the server does not tend to send these messages
very frequently as I have observed disconnected games being left in this
state for hours until I notice and restart gtp2ogs so it will make
another move in the game.

So this change changes game timeouts to only happen when the opponent
has been thinking for longer than the --timeout period. This means we
still disconnect from games where we don't have anything to do but we
are guaranteed to receive an active_game notification whenever there is
something for us to do again (opponent has made a move).

As it happens, this approach is also slightly more self-contained.