bot: track bot failures and exits
Currently when the bot command exits (crashes) in gtp2ogs does not
necessarily notice and may keep on waiting for a response (that will
never arrive).

This change covers a few bot failure modes and makes sure Game can react
to them:

  1. Failure to spawn a process
  2. Failure to write to process stdin
  3. Bot command failing
  4. Bot exiting silently in response to a command

Bot will now set `dead` and `failed` attributes on itself to indicate
the state that it is in so that Game can count the failures and restart
the bot as appropriate.

Currently, Game then proceeds to send a pass to OGS server on failures.
This change replaces that with a disconnect-connect pair so it would not
cause an actual pass move to be played.

In order to break out of infinite loops, Game tracks the number of
retries it has attempted and will resign the game after 5 retries. The
failure count is not persisted across different Game objects referring
to the same underlying game though (say, should we disconnect and stop
tracking the Game).

The change adds synthetic test cases to validate some of this behavior.
I also tested offline with shell scripts that pretend to be broken GTP
clients.