idle-timeout: don't disable idle game timoeut on server disconnect
In f47fc6 (rework game timeouts) I blindly followed the pattern of
disabling idle game timeouts when the server disconnects and disabled
the idle game checker on server timeout.

Only previously the timeouts were per-game and would be reestablished
when the server reconnected. But this time, nothing would come and
re-set the idle game timeout interval.

This commit removes the clearInterval so that idle games are processed
at all times - there will just be nothing to process when the server is
disconnected.

It also adds a testcase for the idle game timeouts. The testcase would
not have caught this bug so it's just for increasing coverage.