only connect to games during bot turn
If there are many games, then connecting to the server can be a noisy
affair if there are many idle correspondence games in flight. We first
connect to all the games and then disconnect from the idle ones.

For bots that have a long list of idle games, that makes the first
connection a needlessly busy event.