game: only kill bot if gamedata has changed
The reason for restarting the bot is that the game state may have
changed and we don't want to implement handling for arbitrary changes
mid-game.

But if the gamedata hasn't changed then there is no reason to reset the
bot status. So let's just keep the bot around.

This may also help address a recent crash loop bug where we keep on
restarting the bot on gamedata, getting a new gamedata and repeating
over and over again.