Pass all notifications as parameters, move functions to bottom of the code, code cleaning and optimizations (#206)
* pass all notifications as parameters and move functions at bottom

- this is standard and makes code much more solid and robust

- added multiple code tweaks all over the code and efficient
code cleaning

- deprecated alias t. for for notification.time_control. :
now that all notifications are passed as parameters at the
bottom of the code, we almost don't use the t. alias anymore
keeping the user. alias for now though

* add new simple functions, various code cleaning and optimizations

- add beforeRankedUnrankedGamesSpecial function :
this new functions allows us to completely remove all the
repetitions in rankedUnranked define at many places of the code

- replace familyArray with familyObject :
as suggested by @Dorus, using an object is clearer and more
practical to access minmax family elements

- remove user.username as a parameter from most functions :
instead of returning in conn_log "user X wanted Y, minmax is Z",
now we simply do "Y is below/above minmax Z", which is enough
because the rejected challenge appends username and user ranking
  The only exception here is bannedFamilyReject function which
still keeps as a parameter user.username, as it is associated
with user.id in messages

- comment cleaning all over the code
- minor comments fixes and tweaks

* improve algorithms efficiency

- using else whenever possible to increase efficiency
- removing repetitive checks on the same condition or its opposite
- for fakerank, we already check earlier in the code if automatic
handicap is not allowed, so no need to double check it in fakerank
code