rework checkChallenge Users/GameSettings to PreRequirements/Settings
* rework checkChallenge Users/GameSettings to PreRequirements/Settings

- the old checkUser function was limiting us regarding the number and
type of checks we could perform : now doing all the pre-requirements
checks first (which if they end up rejecting a challenge, it is
pointless to check game settings such as boardsize or komi or maintime)

- then the old checkGameSettings function is now checkSettings, which
now does not check all game settings, but only checks the lower priority
ones that make sense if challenge is acceptable (checkPreRequirements
are OK)

- reworked maxconnectedgames and maxconnectedgamesperuser checks :
less repetitions, tweaked algorithm, also since we provide defaults
for config.maxconnectedgames and config.maxconnectedgamesperuser
we dont need to check these args
- using const whenver possible now
- minor tweaks and fixes as well
- now that we pass notifications as parameters, we almost
don't use the old alias user. anymore, so we can deprecate it now
- this is also a consistency change with other parts of the code
where we always use notification.user
- also, for the bans family, we will replace the current code with
a refactored function that will only need once the notification.user
and notification.id, so this alias removal is an anticipation to
that future change

* add a check for arg families that dont come with a default

- for example config.minhandicap may be false if bot admin doesnt
choose a value, since we provide no defaults for handicap number.
Hence we have to check if arg

* clearer naming

- as @Dorus noted, in some languages objectString may confused with
another type than the returned String here in javascript