show jwt only once, then hide it with "already showed once"
as @Dorus explained on discord, jwt is a token we get after we log in,
and it contains information about our login. It is not necessarily
something we need to keep secret, so showing it only once is enough
for debugging purposes, then we can hide it with a "already showed once"
to make logs clearer and more readable

also, as @Dorus said too, it's just a message that say you're logged in
and what your access rights are and some other user info, but the thing
about it is the server signed it so anyone who read it can verify the
server approved this info.