Fix unpause callback
The `GobanEngineConfig.clock` object does not contain `pause` member
for most clock updates, it only exists when pausing or unpausing the
clock. So releasing the unpause callback on every clock update results
results in us never unpausing if there is any clock update, like when
making a move.

We still won't now to unpause the clock if the clock was already paused
when we started up, but I'm leaving that for now.