unpause game if opponent makes a move
I'm happy to allow the opponent take the time to step away from the game
if they need to, but I want to be able to understand how many
blitz/rapid/live games there are and enforce the limits of how many the
bot can handle at the same time.

I think this makes sense in general, but alternatively could also make
it configurable.